move collision logic to player store and fix jump
